LayoutTests/media/track/vtt-cue-root-selector-expected.txt


EVENT(canplaythrough)
EVENT(seeked)
Test that ::cue(:root) targets the cue element's text color.
EXPECTED (cueStyle.color == 'rgb(0, 255, 0)') OK

Test that the document :root selector does not affect the cue element background color.
EXPECTED (cueStyle.backgroundColor != 'rgb(255, 0, 0)' == 'true') OK
END OF TEST
